Create Template Builder
I have created a javascript based template builder
the user submits the width and height
a table then pops up and the user clicks the cells they want to be
alive.
the input is then converted back to a string of xxxooooxxxx's  and saved
in the database.
still need to finnish the error checking for Name input