fixed the gameof life functions
There where a couple of issues that I did not see until I tried a simple
blinker.
Template was copying over itself in reverse on start up, Removed code.

The take turn was not creating a new cell [][] to copy into, thiss means
that it changed the result for the next cell being tested.

The rules where not checking if the cell was already alive or dead
before doing the count.... There are two rules, check out assignment 1
again.